+++ lib/TokenManager.js
@@ -14,7 +14,7 @@
 /* eslint complexity:[0,8] max-statements:[0,21] */
 var util = require('./util');
 var AuthSdkError = require('./errors/AuthSdkError');
-var storageUtil = require('./browserStorageUtil');
+var storageUtil = require('./browser/browserStorageUtil');
 var Q = require('q');
 var Emitter = require('tiny-emitter');
 var config = require('./config');
