+++ lib/browser/browser.js
@@ -12,45 +12,25 @@
 /* eslint-disable complexity */
 /* eslint-disable max-statements */
 
-require('./vendor/polyfills');
+require('../vendor/polyfills');
 
 var Q                 = require('q');
-var oauthUtil         = require('./oauthUtil');
-var util              = require('./util');
-var tx                = require('./tx');
-var session           = require('./session');
+var oauthUtil         = require('../oauthUtil');
+var util              = require('../util');
+var builderUtil       = require('../builderUtil');
+var tx                = require('../tx');
+var session           = require('../session');
 var cookies           = require('./cookies');
-var token             = require('./token');
-var AuthSdkError      = require('./errors/AuthSdkError');
-var config            = require('./config');
-var TokenManager      = require('./TokenManager');
-var http              = require('./http');
+var token             = require('../token');
+var AuthSdkError      = require('../errors/AuthSdkError');
+var config            = require('../config');
+var TokenManager      = require('../TokenManager');
+var http              = require('../http');
 
 function OktaAuthBuilder(args) {
   var sdk = this;
 
-  if (!args) {
-    throw new AuthSdkError('No arguments passed to constructor. ' +
-      'Required usage: new OktaAuth(args)');
-  }
-
-  var url = args.url;
-  if (!url) {
-    var isUrlRegex = new RegExp('^http?s?://.+');
-    if (args.issuer && isUrlRegex.test(args.issuer)) {
-      // Infer the URL from the issuer URL, omitting the /oauth2/{authServerId}
-      url = args.issuer.split('/oauth2/')[0];
-    } else {
-      throw new AuthSdkError('No url passed to constructor. ' +
-      'Required usage: new OktaAuth({url: "https://{yourOktaDomain}.com"})');
-    }
-  }
-
-  if (url.indexOf('-admin.') !== -1) {
-    throw new AuthSdkError('URL passed to constructor contains "-admin" in subdomain. ' +
-      'Required usage: new OktaAuth({url: "https://{yourOktaDomain}.com})');
-  }
-
+  var url = builderUtil.validateConfig(args);
   this.options = {
     url: util.removeTrailingSlash(url),
     clientId: args.clientId,
@@ -272,19 +252,4 @@ proto.fingerprint = function(options) {
   });
 };
 
-module.exports = function(ajaxRequest) {
-  function OktaAuth(args) {
-    if (!(this instanceof OktaAuth)) {
-      return new OktaAuth(args);
-    }
-
-    if (args && !args.ajaxRequest) {
-      args.ajaxRequest = ajaxRequest;
-    }
-    util.bind(OktaAuthBuilder, this)(args);
-  }
-  OktaAuth.prototype = OktaAuthBuilder.prototype;
-  OktaAuth.prototype.constructor = OktaAuth;
-
-  return OktaAuth;
-};
+module.exports = builderUtil.buildOktaAuth(OktaAuthBuilder, true);
