+++ lib/browser/browserIndex.js
@@ -12,4 +12,4 @@
  */
 
 // This exists to use reqwest for http requests by default when library used on browser
-module.exports = require('../reqwest');
+module.exports = require('../../reqwest');
