+++ lib/browser/browserStorageUtil.js
@@ -12,8 +12,8 @@
  */
 
 var cookies = require('./cookies');
-var storageBuilder = require('./storageBuilder');
-var config = require('./config');
+var storageBuilder = require('../storageBuilder');
+var config = require('../config');
 
 // Building this as an object allows us to mock the functions in our tests
 var storageUtil = {};
@@ -62,7 +62,7 @@ storageUtil.getCookieStorage = function() {
     getItem: cookies.get,
     setItem: function(key, value) {
       // Cookie shouldn't expire
-      cookies.set(key, value, '2038-01-19T03:14:07.000Z');
+      cookies.set(key, value, '2200-01-01T00:00:00.000Z');
     }
   };
 };
