+++ lib/browser/cookies.js
