+++ lib/http.js
@@ -13,8 +13,8 @@
 
 /* eslint-disable complexity */
 var util = require('./util');
-var storage = require('./serverStorage');
-var storageUtil = require('./serverStorageUtil');
+var storage = require('./server/serverStorage');
+var storageUtil = require('./server/serverStorageUtil');
 var Q = require('q');
 var AuthApiError = require('./errors/AuthApiError');
 var config = require('./config');
@@ -55,7 +55,7 @@ function httpRequest(sdk, options) {
   var err, res;
   return new Q(sdk.options.request(method, url, ajaxOptions))
     .then(function(resp) {
-      res = resp.responseText || resp;
+      res = resp.responseText;
       if (res && util.isString(res)) {
         res = JSON.parse(res);
       }
@@ -80,7 +80,7 @@ function httpRequest(sdk, options) {
       return res;
     })
     .fail(function(resp) {
-      var serverErr = resp.responseText || resp || {};
+      var serverErr = resp.responseText || {};
       if (util.isString(serverErr)) {
         try {
           serverErr = JSON.parse(serverErr);
