+++ lib/oauthUtil.js
@@ -14,7 +14,7 @@
 /* eslint-disable complexity, max-statements */
 var http = require('./http');
 var util = require('./util');
-var storageUtil = require('./browserStorageUtil');
+var storageUtil = require('./browser/browserStorageUtil');
 var AuthSdkError = require('./errors/AuthSdkError');
 
 var httpCache = storageUtil.getHttpCache();
