+++ lib/server/server.js
@@ -12,33 +12,18 @@
 /* eslint-disable complexity */
 /* eslint-disable max-statements */
 
-require('./vendor/polyfills');
+require('../vendor/polyfills');
 
-var util              = require('./util');
-var tx                = require('./tx');
+var util              = require('../util');
+var builderUtil       = require('../builderUtil');
+var tx                = require('../tx');
 var storage           = require('./serverStorage');
-var AuthSdkError      = require('./errors/AuthSdkError');
-var config            = require('./config');
+var config            = require('../config');
 
 function OktaAuthBuilder(args) {
   var sdk = this;
 
-  if (!args) {
-    throw new AuthSdkError('No arguments passed to constructor. ' +
-      'Required usage: new OktaAuth(args)');
-  }
-
-  var url = args.url;
-  if (!url) {
-    throw new AuthSdkError('No url passed to constructor. ' +
-    'Required usage: new OktaAuth({url: "https://{yourOktaDomain}.com"})');
-  }
-
-  if (url.indexOf('-admin.') !== -1) {
-    throw new AuthSdkError('URL passed to constructor contains "-admin" in subdomain. ' +
-      'Required usage: new OktaAuth({url: "https://{yourOktaDomain}.com})');
-  }
-
+  var url = builderUtil.validateConfig(args);
   this.options = {
     url: util.removeTrailingSlash(url),
     redirectUri: args.redirectUri,
@@ -81,19 +66,4 @@ proto.verifyRecoveryToken = function (opts) {
   return tx.postToTransaction(this, '/api/v1/authn/recovery/token', opts);
 };
 
-module.exports = function(request) {
-  function OktaAuth(args) {
-    if (!(this instanceof OktaAuth)) {
-      return new OktaAuth(args);
-    }
-
-    if (args && !args.request) {
-      args.request = request;
-    }
-    util.bind(OktaAuthBuilder, this)(args);
-  }
-  OktaAuth.prototype = OktaAuthBuilder.prototype;
-  OktaAuth.prototype.constructor = OktaAuth;
-
-  return OktaAuth;
-};
+module.exports = builderUtil.buildOktaAuth(OktaAuthBuilder, false);
