+++ lib/server/serverIndex.js
@@ -10,32 +10,6 @@
  * See the License for the specific language governing permissions and limitations under the License.
  *
  */
-var NodeCache = require('node-cache');
-var storage = new NodeCache();
 
-function set(name, value, expiresAt) {
-  // eslint-disable-next-line no-extra-boolean-cast
-  if (!!(Date.parse(expiresAt))) {
-    // Time to expiration in seconds
-    var ttl = (Date.parse(expiresAt) - Date.now()) / 1000;
-    storage.set(name, value, ttl);
-  } else {
-    storage.set(name, value);
-  }
-
-  return get(name);
-}
-
-function get(name) {
-  return storage.get(name);
-}
-
-function del(name) {
-  return storage.del(name);
-}
-
-module.exports = {
-  set: set,
-  get: get,
-  delete: del
-};
+// This exists to use fetch for http requests by default when library used on server side
+module.exports = require('../../fetch');
