+++ lib/server/serverStorage.js
@@ -10,25 +10,32 @@
  * See the License for the specific language governing permissions and limitations under the License.
  *
  */
+var NodeCache = require('node-cache');
+var storage = new NodeCache();
 
-var storage = require('./serverStorage');
-var storageBuilder = require('./storageBuilder');
-var config = require('./config');
+function setToCache(name, value, expiresAt) {
+  // eslint-disable-next-line no-extra-boolean-cast
+  if (!!(Date.parse(expiresAt))) {
+    // Time to expiration in seconds
+    var ttl = (Date.parse(expiresAt) - Date.now()) / 1000;
+    storage.set(name, value, ttl);
+  } else {
+    storage.set(name, value);
+  }
 
-// Building this as an object allows us to mock the functions in our tests
-var storageUtil = {};
+  return getFromCache(name);
+}
 
-storageUtil.getHttpCache = function() {
-    return storageBuilder(storageUtil.getStorage(), config.CACHE_STORAGE_NAME);
+function getFromCache(name) {
+  return storage.get(name);
+}
+
+function deleteFromCache(name) {
+  return storage.del(name);
+}
+
+module.exports = {
+  set: setToCache,
+  get: getFromCache,
+  delete: deleteFromCache
 };
-
-storageUtil.getStorage = function() {
-  return {
-    getItem: storage.get,
-    setItem: function(key, value) {
-      storage.set(key, value, '2038-01-19T03:14:07.000Z');
-    }
-  };
-};
-
-module.exports = storageUtil;
