+++ lib/server/serverStorageUtil.js
@@ -11,5 +11,24 @@
  *
  */
 
-// This exists to use fetch for http requests by default when library used on server side
-module.exports = require('../fetch');
+var storage = require('./serverStorage');
+var storageBuilder = require('../storageBuilder');
+var config = require('../config');
+
+// Building this as an object allows us to mock the functions in our tests
+var storageUtil = {};
+
+storageUtil.getHttpCache = function() {
+    return storageBuilder(storageUtil.getStorage(), config.CACHE_STORAGE_NAME);
+};
+
+storageUtil.getStorage = function() {
+  return {
+    getItem: storage.get,
+    setItem: function(key, value) {
+      storage.set(key, value, '2200-01-01T00:00:00.000Z');
+    }
+  };
+};
+
+module.exports = storageUtil;
