+++ lib/token.js
@@ -20,7 +20,7 @@ var sdkCrypto     = require('./crypto');
 var AuthSdkError  = require('./errors/AuthSdkError');
 var OAuthError    = require('./errors/OAuthError');
 var config        = require('./config');
-var cookies       = require('./cookies');
+var cookies       = require('./browser/cookies');
 
 function decodeToken(token) {
   var jwt = token.split('.');
