+++ test/util/oauthUtil.js
@@ -7,7 +7,7 @@ var EventEmitter = require('tiny-emitter');
 var wellKnown = require('../xhr/well-known');
 var wellKnownSharedResource = require('../xhr/well-known-shared-resource');
 var keys = require('../xhr/keys');
-var storageUtil = require('../../lib/browserStorageUtil');
+var storageUtil = require('../../lib/browser/browserStorageUtil');
 
 var oauthUtil = {};
